Update statuses service
