+++ instances/wearApp/src/main/AndroidManifest.xml
@@ -46,19 +46,6 @@
                 android:name="androidx.wear.tiles.PREVIEW"
                 android:resource="@mipmap/ic_launcher" />
         </service>
-
-        <service
-            android:name=".wear.messenger.ping.service.PingListenerService"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="com.google.android.gms.wearable.MESSAGE_RECEIVED" />
-
-                <data
-                    android:host="*"
-                    android:pathPrefix="/wear/ping"
-                    android:scheme="wear" />
-            </intent-filter>
-        </service>
         <service
             android:name=".wear.tile.ComposeTileService"
             android:exported="true"
