+++ modules/services/wear-messenger/ping-wear/src/androidMain/AndroidManifest.xml
