+++ modules/services/wear-messenger/ping-wear/src/commonMain/kotlin/com/makeevrserg/empireprojekt/mobile/wear/messenger/ping/util/PingWearMessage.kt
